Adds GetReservationCountPerMonthAsync method.

+++ Library-Web-Application-NET.Server/src/statistics/MonthCount.cs
@@ -2,7 +2,7 @@
 {
     public class MonthCount
     {
-        public short Month { get; set; }
+        public int Month { get; set; }
 
         public long Count { get; set; }
     }
